+++ plume/src/main/kotlin/io/github/plume/oss/graph/CallGraphBuilder.kt
@@ -18,7 +18,8 @@ package io.github.plume.oss.graph
 import io.github.plume.oss.Extractor
 import io.github.plume.oss.Extractor.Companion.getSootAssociation
 import io.github.plume.oss.drivers.IDriver
-import io.github.plume.oss.util.SootToPlumeUtil
+import io.github.plume.oss.metrics.ExtractorTimeKey
+import io.github.plume.oss.metrics.PlumeTimer
 import io.github.plume.oss.util.SootToPlumeUtil.constructPhantom
 import io.shiftleft.codepropertygraph.generated.EdgeTypes.CALL
 import io.shiftleft.codepropertygraph.generated.NodeTypes.METHOD
@@ -75,6 +76,7 @@ class CallGraphBuilder(private val driver: IDriver) : IGraphBuilder {
             is InvokeStmt -> getSootAssociation(unit.invokeExpr)?.filterIsInstance<NewCallBuilder>()?.firstOrNull()
             else -> null
         }?.let { callV ->
+            PlumeTimer.startTimerOn(ExtractorTimeKey.DATABASE_READ)
             driver.getNeighbours(callV).use { g ->
                 // If there is no outgoing call edge from this call, then we should attempt to find it's target method
                 if (g.node(callV.id())?.outE(CALL)?.hasNext() != true) {
@@ -91,6 +93,7 @@ class CallGraphBuilder(private val driver: IDriver) : IGraphBuilder {
                     }
                 }
             }
+            PlumeTimer.stopTimerOn(ExtractorTimeKey.DATABASE_READ)
         }
     }